+++ src/models/public-models.ts
@@ -245,3 +245,68 @@ export interface IZoneerderServiceConfig {
   agivGrbUrl: string;
   beschermingenWMSUrl: string;
 }
+
+export interface ICrabAdres {
+  gemeente: IGemeente;
+  land: ILand;
+  postcode: IPostcode;
+  straat: IStraat;
+  adres: IAdresregisterAdres;
+}
+
+export interface ILand {
+  code: string;
+  naam: string;
+}
+
+export interface IGemeente {
+  id?: number;
+  naam: string;
+  niscode: string;
+  provincie?: IProvincie;
+}
+
+export interface IPostcode {
+  nummer: string;
+  uri: string;
+}
+
+export interface IStraat {
+  id: string;
+  naam: string;
+  uri: string;
+  omschrijving: string;
+}
+
+export interface IAdresregisterAdres {
+  id?: string;
+  huisnummer?: string;
+  busnummer?: string;
+  uri?: string;
+}
+
+export interface IProvincie {
+  niscode: string;
+  naam: string;
+  gewest: IGewest;
+}
+
+export interface IGewest {
+  naam: string;
+  niscode: string;
+}
+
+export interface IPostinfo {
+  namen: string[];
+  postcode: string;
+  status: string;
+  uri: string;
+}
+
+export interface IActorAdres {
+  gemeente: IGemeente;
+  land: ILand;
+  postcode: IPostcode;
+  straat: IStraat;
+  adres: IAdresregisterAdres;
+}
